+++ crates/scouter_drift/src/psi/monitor.rs
@@ -6,8 +6,7 @@ use num_traits::{Float, FromPrimitive};
 use rayon::prelude::*;
 use scouter_error::MonitorError;
 use scouter_types::psi::{
-    Bin, PsiDriftConfig, PsiDriftMap, PsiDriftProfile, PsiFeatureDriftProfile,
-};
+    Bin, PsiDriftConfig, PsiDriftMap, PsiDriftProfile, PsiFeatureDriftProfile};
 use std::collections::HashMap;
 
 #[derive(Default)]
